move code from index to plugin
